Add paragraph acknowledging right to request and delete personal information.
Technically this is only a legal right in the EU. However, we're happy to give it to anyone (and we have always done so).

The text is exactly what BBB suggested eoxcept that I've changed "Sandstorm" to "We" to be consistent with the rest of the document. "Sandstorm" is ambiguous -- it could mean the company, the service, or the software.