Mention that you can opt out of third-party data sharing by avoiding using features that integrate with third parties.
The BBB suggested text saying that we will provide an explicit opt-out or opt-in, but I found that text confusing. E.g. you can't log in with Github and then opt-out of sharing with Github -- there's no way for us to implement that. It's up to the user to avoid Github login if they don't want any data shared with Github.