U.S.-Swiss Safe Harbor no longer exists.
We've been informed that we must remove this section, as it is obsolete.

The program has been replaced by the U.S.-Swiss Privacy Shield. However, registering for Privacy Shield requires a new set of paperwork and fees. As Sandstorm is no longer operating as a startup, we lack the resources to complete this task. Therefore, we are simply removing the text altogether. (We recently registered for the similar U.S.-E.U. Privacy Shield, and found it to be too much work.)

We nevertheless believe that we are in compliance with all rules in question. We simply lack the ability to do the paperwork.