Our US-EU Privacy Shield registration is expiring and we lack the resources (time and money) to renew it.
Of course, nothing about our actual privacy practices is changing, and we believe ourselves to be in full compliance. We just aren't filing our self-certification nor paying at outside agency to resolve disputes anymore (to date there has never been such a dispute).

Note that we have no operations in the EU; our servers are located strictly in the US.